Cleaning things

+++ game/food.h
@@ -2,18 +2,19 @@
 
 #include <SDL2/SDL_rect.h>
 #include <SDL2/SDL_render.h>
+#include <cstdint>
 #include "../helpers/constants.h"
 
 namespace Foods {
   class Food{
     public:
       Food();
-      unsigned int get_x();
-      unsigned int get_y();
+      int16_t get_x();
+      int16_t get_y();
       void render(SDL_Renderer* render);
       ~Food();
       void update_position();
-      void set_position(unsigned int x, unsigned int y);
+      void set_position(uint16_t x, uint16_t y);
 
     private:
       SDL_Rect *shape = new SDL_Rect{0, 0, FOOD_W, FOOD_H};
